day, event and week components refactored
