+++ src/components/Cal2/Calen.js
@@ -1,8 +1,10 @@
-import React from 'react'
-import moment from 'moment'
-import './Calen.css'
-import axios from 'axios'
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group'
+import React from 'react';
+import moment from 'moment';
+import './Calen.css';
+import axios from 'axios';
+import Week from './CalComponents/Week.js';
+import DayNames from './CalComponents/DayNames.js';
+import Events from './CalComponents/Events.js';
 
 export default class Calendar extends React.Component {
   constructor(props) {
@@ -168,10 +170,11 @@ export default class Calendar extends React.Component {
       if (window.confirm("Are you sure you want to add an event in the past?")) {
         this.handleAdd();
       } else {
-      } // end confirm past
-    } //end is in the past
+        // TODO: else...
+      }
+    }
   }
-//not sure we need this part dont eant to allow deletion through user
+
   removeEvent(i) {
     const monthEvents = this.state.selectedMonthEvents.slice();
     
@@ -189,7 +192,6 @@ export default class Calendar extends React.Component {
       });
     }
   }
-//perhaps this is the axios call that maps our events
 
 componentDidMount(){
   axios
@@ -277,142 +279,5 @@ initialiseEvents() {
         </section>
       );
     }
-  }
-}
-
-class Events extends React.Component {
-  render() {
-    const currentSelectedDay = this.props.selectedDay;
-    const monthEvents = this.props.selectedMonthEvents;
-    const removeEvent = this.props.removeEvent;
-    let rendTemp = moment(monthEvents)
-
-    const monthEventsRendered = rendTemp._i.map((event, i) => {
-      let events = moment(event)
-      return (
-       
-        <div
-          key={events.event_id}
-          className="event-container"
-          onClick={() => removeEvent(i)}
-        >
-          <ReactCSSTransitionGroup
-            component="div"
-            className="animated-time"
-            transitionName="time"
-            transitionAppear={true}
-            transitionAppearTimeout={500}
-            transitionEnterTimeout={500}
-            transitionLeaveTimeout={500}
-          >
-            <div className="event-time event-attribute">
-              {event.start_time}
-            </div>
-          </ReactCSSTransitionGroup>
-          <ReactCSSTransitionGroup
-            component="div"
-            className="animated-title"
-            transitionName="title"
-            transitionAppear={true}
-            transitionAppearTimeout={500}
-            transitionEnterTimeout={500}
-            transitionLeaveTimeout={500}
-          >
-            <div className="event-title event-attribute">Summary: {event.summary}, Description: {event.description}</div>
-          </ReactCSSTransitionGroup>
-        </div>
-      );
-    });
-
-    const dayEventsRendered = [];
-    for (var i = 0; i < monthEventsRendered.length; i++) {
-      if (moment(monthEvents[i].start_date).isSame(currentSelectedDay, "day")) {
-        dayEventsRendered.push(monthEventsRendered[i]);
-      }
-    }
-
-    return (
-      <div className="day-events">
-        {dayEventsRendered}
-      </div>
-    );
-  }
-}
-//Displays calendar render
-class DayNames extends React.Component {
-  render() {
-    return (
-      <div className="row days-header">
-        <span className="box day-name">Mon</span>
-        <span className="box day-name">Tue</span>
-        <span className="box day-name">Wed</span>
-        <span className="box day-name">Thu</span>
-        <span className="box day-name">Fri</span>
-        <span className="box day-name">Sat</span>
-        <span className="box day-name">Sun</span>
-      </div>
-    );
-  }
-}
-
-class Week extends React.Component {
-  render() {
-    let days = [];
-    let date = this.props.previousCurrentNextView;
-    let currentMonthView = this.props.currentMonthView;
-    let selected = this.props.selected;
-    let select = this.props.select;
-    let monthEvents = this.props.monthEvents;
-    let tempEvent = moment(monthEvents)
-    for (var i = 0; i < 7; i++) {
-      var dayHasEvents = false;
-
-      for (var j = 0; j < tempEvent.length; j++) {
-        if (monthEvents[j].date.isSame(date, "day")) {
-          dayHasEvents = true;
-        }
-      }
-
-      let day = {
-        name: date.format("dd").substring(0, 1),
-        number: date.date(),
-        isCurrentMonth: date.month() === currentMonthView.month(),
-        isToday: date.isSame(new Date(), "day"),
-        date: date,
-        hasEvents: dayHasEvents
-      };
-
-      days.push(<Day day={day} selected={selected} select={select} />);
-      date = date.clone();
-      date.add(1, "d");
-    }
-    return (
-      <div className="row week">
-        {days}
-      </div>
-    );
-  }
-}
-
-class Day extends React.Component {
-  render() {
-    let day = this.props.day;
-    let selected = this.props.selected;
-    let select = this.props.select;
-
-    return (
-      <div
-        className={
-          "day" +
-          (day.isToday ? " today" : "") +
-          (day.isCurrentMonth ? "" : " different-month") +
-          (day.date.isSame(selected) ? " selected" : "") +
-          (day.hasEvents ? " has-events" : "")
-        }
-        onClick={() => select(day)}
-      >
-        <div className="day-number">{day.number}</div>
-      </div>
-    );
   }
 }
